[FEATURE] Indexed Search: add mySQL fulltext support
Adds some minor modifications for restructuring in the pi code,
and some hooks in order to allow for other search engines to be
used.
Additionally, a new extension "indexed_search_mysql" is added
to the core that allows the mysql fulltext search with it.
Thanks to Michael Stucki for all the work!

Change-Id: I7e739baa6d35b6678d008f19737f86c3bdaa6dd5
Resolves: #28613
Reviewed-on: http://review.typo3.org/6657
Reviewed-by: Steffen Ritter
Tested-by: Steffen Ritter