[BUGFIX] PageBrowsing ViewHelper defines unused method argument
The PHP-Doc comments for the render-method in PageBrowsing 
Viewhelper on Indexed Search define a $details argument 
for that method with is then not part of the method arguments 
itself.

As it seems to be unused the argument definition should simply 
be removed from doc comments to dont mess up the defintions 
of the other arguments.

Right now this results in maximumNumberOfResultPages be defined 
as array and causes a fatal error like:
The argument "maximumNumberOfResultPages" was registered with 
type "array", but is of type "string" in view helper 
"TYPO3\CMS\IndexedSearch\ViewHelpers\PageBrowsingViewHelper"

Resolves: #54807
Releases: 6.2, 6.1
Change-Id: I58376880ef64a1e1088561fbdb3ff25f9c1fbdc2
Reviewed-on: https://review.typo3.org/26676
Reviewed-by: Anja Leichsenring
Tested-by: Anja Leichsenring