[CLEANUP] Cleanup jumpToUrl() and unused JS in Backend
The jumpToUrl() is defined in every
second module as inline JS. As the
duplicated code is erroneous and
can be shimmed completely, the
function is added to every doc-template
code by default, but can be overriden and
removed completely - which is still the case
in some existing modules.

Additionally the variable "script_ended"
is not used at all anymore, and was removed
as well.

Releases: 6.2
Resolves: #56667
Change-Id: I716611bffce402c5f423c3d64c763d6675a79d14
Reviewed-on: https://review.typo3.org/28180
Reviewed-by: Markus Klein
Tested-by: Markus Klein
Reviewed-by: Anja Leichsenring
Tested-by: Anja Leichsenring