[TASK] Use RootlineUtility directly
Instead of calling PageRepository to fetch the rootline, RootlineUtility
can do that now directly.

However, all exceptions are more specific and caught specifically, for
each use-case.

Additionally, PageRepository->getRootline() is deprecated now. For this
reason, the deprecation for the argument $ignoreMPerrors is removed
again, as it does not make sense anymore.

Resolves: #85557
Releases: master
Change-Id: I5a9192311c5616f5a58e89a39c3869b593e97968
Reviewed-on: https://review.typo3.org/57590
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>