[!!!][TASK] Request processing chain must return a response object
This patch introduces a proper request/response handling chain. That
means, that Extbase no longer creates a response object early, passes
it through the request handler and dispatcher into the controller and
expects it to be altered there.

Instead, request handlers, the dispatcher and controllers only get
passed a request object and controllers are then responsible for
creating and returning a response.

This change is preparatory work for the introduction of PSR-7 responses
in Extbase. Those responses are immutable and can no longer passed by
reference through the stack. Instead, the controller must return the
desired response along with its state down the stack while remaining
unaltered.

This patch is considered breaking because it changes several method
signatures of public classes and interfaces.

Still, it is quite unlikely to be struck by this change when using
the public api of Extbase and not overriding classes such as the
ActionController.

Releases: master
Resolves: #92513
Change-Id: I3ab7846268ea68727a2f57eea69328b85be4f5e6
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/66079
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Benni Mack <benni@typo3.org>