[TASK] Follow best practice on returning PSR-7 response
Since #94186, extbase controller actions are
required to return a PSR-7 response interface.

To ease the use for extension authors, the
helper method `htmlResponse()` was added
to the ActionController. If no data is provided,
$this->view->render() is automatically used
as response body.

This is more or less the usual behaviour of
most extbase actions and therefore also
described in the corresponding RST.

Since core code is often used as example
for custom implementations, those places
should follow the best practices, described
in the RST.

Therefore, this patch adjust all places in core,
which call htmlResponse() and unnecessarily
pass $this->view->render() as data.

Resolves: #94186
Releases: master
Change-Id: Id6ac4f92cc9b6fdb70150a2de4ea5a6877a9a55f
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/69229
Tested-by: core-ci <typo3@b13.com>
Tested-by: Jochen <rothjochen@gmail.com>
Tested-by: Benjamin Franzke <bfr@qbus.de>
Reviewed-by: Jochen <rothjochen@gmail.com>
Reviewed-by: Benjamin Franzke <bfr@qbus.de>