[BUGFIX] Avoid parameter collision in IndexedSearch
The action to delete an indexed item in the IndexedSearch
backend module used the "id" parameter to pass the indexed
item id to the method. The controller is based on extbase and
previously, those parameters were prefixed with the plugin
namespace. However, with #97096 the parameters in extbase
backend modules are no longer prefix by default, which therefore
lead to a collision. The "id" parameter should always represent
the currently selected page, underlying code performs access
checks on this value.

This is now resolved by using a dedicated parameter, which
is not already "reserved" by underlying functionality.

Resolves: #97766
Related: #97096
Releases: main
Change-Id: If7ccb1dfbdad6f907eb4f27187eb7eb9b753e9dc
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/74885
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Bartsch <bo@cedev.de>