[BUGFIX] Change column metaphone to varchar
In case of activated STRICT_TRANS_TABLES, better known as strict_mode,
the index_words record incl. metaphone can not be stored as a string
is provided, but an int is expected.
As various metaphone related methods in TYPO3 return a string
we change the DB type of column metaphone to varchar. We set
the length to 60 as it is the same we use in analyzeHeaderinfo().

Resolves: #99547
Releases: main, 12.4, 11.5
Change-Id: Ice4223f2169595b781ed8c3a7dd5dda3f50e8a1e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/79817
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>