[BUGFIX] Remove pagination arrows from translations files
We are removing the pagination arrows from the translation
files since they needed to be marked with "aria-hidden" to
avoid acessibility issues. This would extend the usage of
markup that should not  be placed in translation files in the
first place. To allow integrators to change the apperance we
are now adding css classes that can be used to apply
special styling for the next and previous links.

Resolves: #101283
Releases: main, 12.4
Change-Id: I090da2922dd644e161440e4e5a1e47ef0963d839
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/79858
Reviewed-by: Benjamin Kott <benjamin.kott@outlook.com>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benjamin Kott <benjamin.kott@outlook.com>