[BUGFIX] Prevent reaching the placeholder limit in `indexed_search`
The main query execution builds a QueryBuilder instance out of
different sub information. Additionally, several hooks can be
used to influence the created query. On top of this, different
type of search queries are possible.

This change avoid consuming a lot of placeholder for list based
`in()` expressions by using a corresponding QueryBuilder method
suitable for the value type. This means that no placeholder are
used any longer for list expressions.

Note: Max query size may still be exceeded if the lists would
      getting bigger. This can be adjusted by database server
      administrators, but will lead to another exception for
      really large value lists.

Resolves: #86859
Releases: main, 12.4, 11.5
Change-Id: Ic0ac51c13cb74a058d71152ad626c484f696b176
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/79913
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Stefan B�rk <stefan@buerk.tech>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Stefan B�rk <stefan@buerk.tech>