[BUGFIX] Prevent default behavior only for page selector links
The change for #101271 introduced a global event handler to replace the
inline onclick events. The current approach breaks all links on the page
once paging is available as the default click behavior is disabled too
early.

This patch now applies `preventDefault()` only if the clicked link is an
actual `.tx-indexedsearch-page-selector` element.

Resolves: #101331
Related: #101271
Releases: main, 12.4
Change-Id: I0dce23a68555f579942a3daa894b4b0d051ba952
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/79980
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>