[!!!][TASK] Remove indexed_search metaphone search
"Advanced search" Frontend of indexed_search comes
with a "metaphone" / "sounds like" feature.

Its search results are staggeringly bad and the
implementation is a not maintained mess that has
not seen any substantial improvements for 15 years.

The feature seems to be very unused, it seems there
is not even some extension that tries to improve
it, and there are no reports about the incomplete
results one would expect from a proper "metaphone"
implementation. It does not consider UTF-8 in any
way, to name just one issue.

To unblock further indexed_search development that
polish and improve other features, this patch
removes the entire 'metaphone' implementation
together with the Frontend selection, the Backend
statistics, a custom metaphone implementation,
and two database fields.

Resolves: #102900
Releases: main
Change-Id: I2e8f8cb0db3ef75ad83bb1f8189452fd2a2d3a87
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/82600
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Andreas Kienast <a.fernandez@scripting-base.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Andreas Kienast <a.fernandez@scripting-base.de>
Tested-by: Benni Mack <benni@typo3.org>