[!!!][TASK] Remove several outdated Indexed Search options
This change removes various functionality
within Indexed Search which was partly kept
when migrating the plugin towards Extbase
back in TYPO3 v6.2, but serve no real purpose
anymore:

- CSSsuffix
- specialConfiguration
- iconRendering
- flagRendering

This functionality is nowadays represented
within the actual View rendering, thus
the TypoScript options can be removed, as
Fluid is much more powerful than the old
marker-based templating approach.

In addition, the feature to choose a specific
language within the search is removed,
as this functionality is not viable in modern
sites (example: when searching on a french
website, search should not find japanese
results).

Resolves: #102921
Releases: main
Change-Id: I8ff4876f4ad3b8199abb2ff3f0cbf08f382501b7
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/82642
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Andreas Kienast <a.fernandez@scripting-base.de>
Reviewed-by: Andreas Kienast <a.fernandez@scripting-base.de>
Tested-by: Georg Ringer <georg.ringer@gmail.com>