[BUGFIX] Prevent crashes setting stopwords in IS BE module
Argument $pageId is mandatory in saveStopwordsKeywordsAction()
since #102914. The link to that action is created in a fluid
section three times: Once with a proper pages uid, twice without.

This leads to crashes when trying to "Save stopwords" in the
indexed_search BE "detail" module in sections "Words by top count"
and "Words by frequency".

The patch adds the pages-uid argument for these two sections and
simplifies the fluid template a bit along the way.

Change-Id: I3819733be4937b56325e8871dde47b0d00a27742
Resolves: #103543
Related: #102914
Related: #103541
Releases: main
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/83677
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: Garvin Hicking <gh@faktor-e.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>