[BUGFIX] Working link to word details in indexed_search
The indexed_search BE module links to word details.
This link is broken in main and v12:

* In main, the wordHash argument is a string, but the extbase
  target action expects an int.
* In v12, the wordHash is an int, but is misinterpreted as
  pages-uid.

The patch fixes this by switching from the argument name
'id' to 'wordHash' to suppress the pages-uid "convention"
to kick in. In main, the target action argument is declared
as string and both arguments are set non-optional.

Resolves: #103547
Related: #102975
Related: #96797
Related: #103540
Releases: main, 12.4
Change-Id: Ibd4732d086f1af264ee317c7e7d4fe0f15a15aa4
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/83680
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>