[TASK] Remove IS ext_conf_template.txt maxAge
The config label:

"Deprecated! Max TTL (hours) for indexed page: The time in hours
that is the maximum age of an indexed page before it will get indexed
again. Usually the index is kept consistent, therefore you won't need
this feature."

This message is there since ... *doing some git archaeology* ...
SVN id 2006 in year 2007! That's the Changelog entry:

"Turn off maxAge setting as this "feature" seems to make no sense
at all."

Agreed: When content is changed in BE, an IS hook removes the
page index, and it will be re-indexed with next call (or via
crawler). It seems it's finally time to remove the option plus
the attached code without further notice ;)

Change-Id: Ia8031ca4bf58d7c8293c1b8eb3f672a1fac2266b
Resolves: #103548
Releases: main
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/83683
Tested-by: Andreas Kienast <a.fernandez@scripting-base.de>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Andreas Kienast <a.fernandez@scripting-base.de>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>