[TASK] Avoid indexed_search lexer debug
The indexed_search class Lexer is used to split strings
into words using some weird multibyte magic.

It has a "debug" mode, enabled during indexing when
ext_conf_template.txt debugMode is true. It then creates
a HTML string in its main split2Words() method, which is
later written to index_debug along with other
debug information. That debug string is then output in
the BE module when looking at indexing details, if given.

The fun part is, this debug string is reset in-between
multiple split2Words() calls, and any indexing always calls
the method four times, with the last one being 'description',
which is usually empty.

This means there is in practice almost never any debug
output to show in BE, and even if shown, it is only a
fraction of the processed content.

The patch removes this lexer debug handling.

Resolves: #103550
Releases: main
Change-Id: I406a621bc88fc320f348f2e4ba2cf7bc0882e421
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/83687
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Andreas Kienast <a.fernandez@scripting-base.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Andreas Kienast <a.fernandez@scripting-base.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>