[TASK] Avoid managing keywords in IS BE module
The "Indexing" indexed_search backend module allows
picking single indexed words and adding them as
keywords a pages 'keywords' field.

This mixes up two different topics: Search and seo.

Good seo administration should be a separate topic that
involves much more than just adding some arbitrary
keyword to a page. Additionally, meta keywords lost quite
a bit of relevance over the years, and it seems odd the
indexed_search BE module takes care of this detail. It
additionally comes with zero additional explanation for
editors, is hidden in detail views of the module, and
removes the index of a page when used, which leads to
unexpected routing within the module.

The patch removes this functionality.

This also fixes an issue when trying to save keywords
for an index that is not a page, but something else
like an image.

Resolves: #103657
Resolves: #103618
Releases: main
Change-Id: Ifd0e1de9c5d02082e2abb363dcd435ecb6778337
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/83878
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>