[TASK] Remove "Access for 'admin' users only!" hints in About module
The "About" module lists, among other things, all backend modules
available in the system. If an admin or editor does not have access to
a module, it will not be displayed in the "About" module list.

Some of the modules listed contained a note that the module was only
accessible to administrators. With the introduction of the system
maintainer role, this note is no longer applicable. In addition, this
note was missing for some modules that cannot be made accessible to the
editor at all (e.g. "Access" module).

If an admin/system maintainer wants to check which backend modules an
editor can access, he usually switches to the backend user (via the
"Backend Users" backend module). Calling the "About" module is not
sufficient at this point, as settings in backend user groups and the
backend user accounts themselves also influence whether a backend user
can access a backend module or not.

For this reason, this note in the "About" module is now being removed.

Resolves: #104040
Releases: main, 12.4
Change-Id: Ib27581a9f1fc48985d42550622caffcf0615a0d4
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/84610
Reviewed-by: Michael Telgkamp <michael.telgkamp@mindscreen.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Jasmina Ließmann <minapokhalo+typo3@gmail.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Michael Telgkamp <michael.telgkamp@mindscreen.de>
Tested-by: Jasmina Ließmann <minapokhalo+typo3@gmail.com>