[TASK] Allow translation of site sets and settings labels
The set label and setting labels and descriptions are now
translatable. That means the `label:` and `description:`
keys may be ommited and and will be read from a labels-file
by convention:

 * set label: `label`
 * settings label: `settings.{settings-key}`
 * settings description: `settings.description.{settings-key}`

The labels file is referenced centrally via the sets config.yaml file.

Note that the settings labels have not been consumed yet,
implementation will follow in #104794.
TYPO3 documentation hass been adapted to understand this change via:
https://github.com/TYPO3-Documentation/render-guides/pull/734

Resolves: #104831
Related: #104794
Releases: main
Change-Id: I9b0f01891db2d7b0c2db63efcd7cb5408439bd24
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/85869
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: Benni Mack <benni@typo3.org>