[BUGFIX] Remove unwanted spacing of table in "Detailed statistics"
The bottom margin of a table is set to 0 if it's a direct child of
.table-fit.
In the "Detailed statistics" view of Indexing module, due to the
"typo3-indexed-search-statistics" custom element wrapping the table,
this one is not a direct child anymore.

This patch moves the custom element one level upper and adapts the
Javascript accordingly so the "Words found on this page" table can
stay sortable.

Resolves: #104775
Releases: main
Change-Id: I876f39091d4e3229071357b778100052ed2c9a93
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/86249
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: Jasmina Ließmann <minapokhalo+typo3@gmail.com>
Reviewed-by: Guido Schmechel <guido.schmechel@brandung.de>
Reviewed-by: Jasmina Ließmann <minapokhalo+typo3@gmail.com>
Tested-by: Guido Schmechel <guido.schmechel@brandung.de>