[BUGFIX] Show number of total records found in EXT:indexed_search
With the refactorings in #102925 and #102945 the total number
of search results was referred to in `result.pagination.totalAmount`.

However due to the Slice-pagination used, that variable will
not contain the right contents and needs to be replaced with
`result.count`.

This change adapts the default indexedsearch Search.html template
and adjusts the Changelog entry.

Resolves: #104861
Related: #102925
Related: #102945
Releases: main, 13.4
Change-Id: Ia7d8d42d7884f776d996318c75c115722433f564
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/86704
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Mogens Fiebrandt <mogens.fiebrandt@googlemail.com>
Reviewed-by: Mogens Fiebrandt <mogens.fiebrandt@googlemail.com>
Tested-by: core-ci <typo3@b13.com>