[DOCS] Update the Indexing Configuration section
Reduce the number of screenshots to those really needed.
In my eyes one screenshot of the indexing configuration
in general should be enough.

I replaced the screenshots with lists of which values to
set. Those lists are also more accessible and easier to
update in the future.

I moved all these very short pages into just one page
and condensed and shortened the texts.

I tested the configuration and removed parts that
no longer worked.

A more detailed description of how the crawler extension
works needs to be described in the extension's manual
and has been removed from here with references to the
extension. (I had to use an absolute link to the
manual as something seems to be broken with that
manual interlinks, will investigate).

Resolves: #106137
Releases: main, 13.4
Change-Id: I4af2b90c0db42f30e18e7a6e746205ecf2ab2a79
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/88119
Reviewed-by: Stefan Froemken <froemken@gmail.com>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>