[TASK] Declare private properties in final classes
The patch declares the majority of properties and
constants of final classes private, partially methods
as well. Some more edgy cases are left out for now.
Doing this is technically kinda irrelevant but can
help IDE's to find unused properties and methods more
easily. phpstan is also a bit more picky, a couple of
entries are added and coupled with some @todo comments
in the codebase. The patch comes with minor adaptions
like slightly improved DI and similar.

Resolves: #107965
Releases: main
Change-Id: Ic0f064fe9aaed029176f098600e4fb0e36b7552a
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/91416
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>