[BUGFIX] Remove wrong condition for re-indexing
This bugfix is affected by two previous patches.
The first one #102955 migrated the IndexStatus to an enum,
which was previously represented by numbers ranging
from negative to positive values. Positive values
meant "re-index required". So all calls to "> 0" were
migrated to check for the new boolean value instead.
The one affected condition however was checking for
"=== 1". The value "1" was representing "maxAge exceeded".
This status was already removed in #103548. This means
we can safely remove the second part of the condition.

Resolves: #108251
Related: #102955
Related: #103548
Releases: main, 13.4
Change-Id: I646f9e6bc68ed4b948d45622a23089c736f85f1e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/91811
Tested-by: Jasmina Ließmann <typo3@minapok.eu>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Jasmina Ließmann <typo3@minapok.eu>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>